refactor: Generize Shadows, and move Properties to Shadows
Properties -> Shadows: before we had a "properties" prop in our fields, in fact it isn't true because the project is "shadow" oriented, so we needed to rename "properties" to "shadows" to match better the field's goal. We would add in future "otherProperties" next to "shadows" to handle every other wanted properties that aren't shadows. Also generize "Shadows", in case where we want to consider any other property as a shadow in futur.

feature/sc-95