fix: Fix splitted field with errors
If we split a field which contains errors, it will patch the hovered field with concerned default values. It requires to know which is the hovered field before splitting it, so we add the getHoveredField method that returns it. Since we can retrieve the hovered field before splitting it, we should now pass it to the split function.

feature/sc-100