fix: Remove keys in editor/help-keyframes which forces components rerender
The main problem was that every curry functions for fields were not reactive to their props, leading to modify the object with which they were initialised at component setup. Now, we encapsule curry functions in a computed property, permitting us to always have an updated version of the applyUpdater on the new selected object.

bug/sc-121